[JI] Move JsonIndex code to ydb/library (#41374)
