+++ ydb/core/base/ut/ya.make
@@ -25,7 +25,6 @@ SRCS(
     statestorage_guardian_impl_ut.cpp
     statestorage_ut.cpp
     table_index_ut.cpp
-    json_index_ut.cpp
 )
 
 END()
