+++ ydb/core/kqp/opt/logical/kqp_opt_log_json_index.cpp
@@ -2,7 +2,7 @@
 
 #include <expected>
 
-#include <ydb/core/base/json_index.h>
+#include <ydb/library/json_index/json_index.h>
 #include <ydb/core/kqp/common/kqp_yql.h>
 #include <ydb/core/kqp/opt/kqp_opt_impl.h>
 #include <ydb/core/kqp/provider/yql_kikimr_provider_impl.h>
