+++ ydb/core/kqp/opt/logical/ya.make
@@ -19,6 +19,7 @@ SRCS(
 PEERDIR(
     ydb/core/base
     ydb/core/kqp/common
+    ydb/library/json_index
     ydb/core/kqp/opt/cbo
     ydb/core/kqp/opt/cbo/solver
     yql/essentials/core/extract_predicate
