+++ ydb/core/kqp/runtime/kqp_fulltext_analyze.cpp
@@ -2,7 +2,7 @@
 
 #include <ydb/core/base/fulltext.h>
 
-#include <ydb/core/base/json_index.h>
+#include <ydb/library/json_index/json_index.h>
 #include <yql/essentials/minikql/computation/mkql_computation_node_holders.h>
 #include <yql/essentials/minikql/computation/mkql_computation_node_impl.h>
 #include <yql/essentials/minikql/mkql_node_cast.h>
